Fix bug with text on rare anvils
When a rare anvil is single clicked it tries to display the item name
as with a regular anvil but since the text is styled it shows the
markup instead of the name.

Other rare items do not display a name when clicked, so this change
overrides the click functionality to be similar.

Fixes #36