Fixed Tailor BOD not being created intermittently
The leatherworking craft system was always being picked, regardless of material used. That meant if cloth was picked (50% of the time), it would fail to find an appropriate item that could be crafted, and the gump wouldn't appear as the deed was never created.

Now, the system is picked based on whether or not leather material was randomly rolled.