Allow navigation of craft gump during queue
Previously, the entire gump was rendered useless by blocking all actions (even closing it!) if there was a craft queue active.

Now, the gump can be navigated while waiting for the craft queue to complete.

The changes in this PR allow "non-volatile" buttons to remain responsive, such as viewing categories, picking a material, etc. Now you can actually close the gump instead of waiting for the queue to finish!

Starting a new craft queue is still restricted while waiting for a craft queue to finish.

These changes only apply when S_CraftMany is set to true and have no effect on how quickly the craft queue finishes.