Reduce craft queue timeouts
The excessive timeout on the craft queue finishing and the craft message appearing made for clunky UX by limiting functionality of the gump. It's also confusing that the items appear instantly but then players have to wait 5 seconds to read the results and 7 seconds to utilize the gump again.

This might seem unbalanced but the S_CraftMany setting allows for crafting in batches of 100, so decreases this queue timer makes for a more seamless experience.