Fix issue #364: Remove executables derived from source
These files are all binaries derived from source code in this
repository.

World.exe can be generated as part of the release process for new
releases and included in the artifacts.

The Data.bin/hash/pdb/ref are generated by the server at runtime and so
aren't needed in either location.