Fix issue #369: remove Saves directory from repository
The Saves directory seems to be seeded from some prior run of the shard,
but isn't needed. If it doesn't exist, the user will automatically be
prompted to create an admin account when they first run the server and
then only need to run the [buildworld] command. The [buildworld] command
will use the data from /Data/Spawns to recreate the needed NPCs,
spawners, decorations, etc.

This way there isn't already an admin account with a set password, there
isn't existing statistics, etc.