Automate build and release
This change makes it so that each push will trigger an automatic build
on GitHub's servers. This doesn't cost anything and completes in only a
couple of minutes. If the build succeeds, it will then check to see if
the push was of a tag. If the push is of a tag, it will additionally
produce a new GitHub release with artifacts named after the tag.

This will produce both Windows and Linux builds as well for every
release.

A badge has been added to the top of the README to indicate that the
build is succeeding as of that revision (or not).