Convert the Manual.pdf to a Markdown file
The contents of the manual are wrong (they were wrong before) but I
chose not to fix these issues as part of this change. It will be easier
to fix (and review) these in future commits once this change has landed.

The only changes to the manual text:

- Removed one image that didn't make sense in the context since we can't
  right-align images in GitHub Markdown (the one that just shows the two
  folders for TazUO). Instead, I replaced it with text.
- Some images that were of console text or source code have been
  replaced with code blocks to enable copy/paste.
- A few trivial and obvious typos or grammatical errors were fixed.
- Paths were made consistent (using backticks instead of quotes).

Fixes Secrets-of-Sosaria/World#365